pyproject.toml
[tool.poetry]
name = "airflow-diagrams"
version = "1.0.0.dev0"
description = "Auto-generated Diagrams from Airflow DAGs."
authors = ["Felix Uellendall <feluelle@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
typer = "0.4.0"
diagrams = "0.20.0"
thefuzz = {version = "0.19.0", extras = ["speedup"]}
PyYAML = "6.0"
toml = "0.10.2"
apache-airflow-client = "2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
